refactor: move files & set generic stack propss
