+++ ReactNativeStudy/Ode/Week3/src/components/EmptyComponent.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {Text} from 'react-native';
-import Colors from '../../styles/colors';
+import Colors from '../styles/colors';
 
 export default function EmptyComponent() {
   return <Text style={{color: Colors.White}}>비었어유</Text>;
