+++ ReactNativeStudy/Ode/Week3/src/components/LoadingMoreIndicator.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {Text} from 'react-native';
-import Colors from '../../styles/colors';
+import Colors from '../styles/colors';
 
 export default function LoadingMoreIndicator() {
   return <Text style={{color: Colors.White}}>더 불러오는 중...</Text>;
