+++ ReactNativeStudy/Ode/Week3/src/features/search/components/SearchButton.tsx
@@ -1,10 +1,11 @@
 import React from 'react';
 import {TouchableOpacity} from 'react-native-gesture-handler';
-import SvgIcon from './SvgIcon';
 import {useNavigation} from '@react-navigation/native';
+import SvgIcon from '../../../components/SvgIcon';
+import {ScreenNavigationProp} from '../../../navigators/types';
 
 export default function SearchButton() {
-  const navigation = useNavigation();
+  const navigation = useNavigation<ScreenNavigationProp<'SearchScreen'>>();
   return (
     <TouchableOpacity onPress={() => navigation.navigate('SearchScreen')}>
       <SvgIcon icon={'Search'} />
