+++ ReactNativeStudy/Ode/Week3/src/features/video/types/video.ts
