+++ ReactNativeStudy/Ode/Week3/src/navigators/TabNavigator.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {createBottomTabNavigator} from '@react-navigation/bottom-tabs';
 import Colors from '../styles/colors';
-import SearchButton from '../components/common/SearchButton';
+import SearchButton from '../features/search/components/SearchButton';
 import SvgIcon from '../components/common/SvgIcon';
 import commonHeaderScreenOptions from '../styles/commonHeaderScreenOptions';
 import HomeScreen from '../screens/HomeScreen';
